1. fixes bugs

+++ app/src/main/java/com/ebabu/event365live/listener/SelectedVipTicketListener.java
@@ -1,5 +1,5 @@
 package com.ebabu.event365live.listener;
 
 public interface SelectedVipTicketListener {
-    void getSelectedTicketListener(int ticketId, String ticketType,  int ticketPrice,int ticketQty);
+    void getSelectedTicketListener(int ticketId, String ticketType,  double ticketPrice,int ticketQty);
 }
