+++ app/src/main/java/com/ebabu/event365live/ticketbuy/modal/CalculateEventPriceModal.kt
@@ -1,4 +1,4 @@
 package com.ebabu.event365live.ticketbuy.modal
 
-data class CalculateEventPriceModal(var ticketId: Int, var ticketType: String, var ticketPrice: Int, var ticketQty: Int) {
+data class CalculateEventPriceModal(var ticketId: Int, var ticketType: String, var ticketPrice: Double, var ticketQty: Int) {
 }
